This commit cleans code up by making it component based
